+++ axiom-macros/tests/compile_fail/missing_name.rs
@@ -1,5 +1,0 @@
-// Copyright (c) 2026 Kirky.X
-use axiom_macros::service_api;
-
-#[service_api(version = "v1")]
-async fn test_fn() {}